==> notebooks/3_1_AddDerivedAttributes.ipynb <==
# 3.1 Adding Derived Attributes to existing Data #
## For Brownlow Predictor Project ##

Adds Derived Attributes of 
- [x] Uncontested Marks
- [x] Marks Outside 50
- [x] Tackles Outside 50
- [x] Behind Assists
- [x] Effective Disposals
- [x] Ineffective Disposals

and deleted
- [ ] Disposal Efficiency %

from data

*Completes the data picture in terms of all the 'Disposals/Kicks/Handball' triangle relationship, and resolves the problem of having a column of data being a percentage*

**Author: `Lang (Ron) Chen` 2021.12-2022.1**

---

**0. Import Libraries**

import pandas as pd
import os

choice = 'OriginalData'

**1. Opens each file and performs operation**

# Grabs a list of the files
filelist = os.listdir(f'../data/raw/{choice}')
# Remove the first file (an ipynb checkpoint file)
filelist.sort()

%time

# Open all the files, performs operations and exports them to a new folder

if not os.path.exists(f'../future data/curated/OriginalData_AddDerived'):
    os.makedirs(f'../future data/curated/OriginalData_AddDerived')

for file in filelist:
    if file[0] == '.':
        continue
    
    df = pd.read_csv(f'../data/raw/{choice}/{file}')
    
    df['Uncontested Marks'] = df['Marks'].sub(df['Contested Marks'])
    df['Marks Outside 50'] = df['Marks'].sub(df['Marks Inside 50'])
    df['Tackles Outside 50'] = df['Tackles'].sub(df['Tackles Inside 50'])
    df['Behind Assists'] = df['Score Involvements'].sub(df['Goal Assists'])
    df['Effective Disposals'] = df['Disposals'].mul(df['Disposal Efficiency %']*0.01).round(0)
    df['Ineffective Disposals'] = df['Disposals'].sub(df['Effective Disposals'])
    
    df = df.drop('Disposal Efficiency %', axis = 1)
    
    df.to_csv(f'../data/curated/OriginalData_AddDerived/{file}', index = False)
